The XML can be cut in the last three characters.
Using PHP on Apache2 if a XML response is longer that 8000 chars 3 or 6 last chars are cut. This is produced on printXML method of `utilities/general.php`.

Seems that the problem is due to this PHP bug:
  https://bugs.php.net/bug.php?id=45945
Also comented on:
  http://stackoverflow.com/questions/22121282/php-inserts-hex-number-of

NOTE: Unfortunately I was unable to determine whether possible to lose more characters, but those tests
 I made this patch allows working with Aixada without this problem.