Able to use align on table.tblListingDefault
This change allow to use classes `textAlignRight` & `textAlignCenter` on `<th>` and `<td>` elements without a sub-element as `<p>`.
So allows for a less overhead html code.

This affects some existing presentations that had not been taken into account that alignment was forced to the left on `table.tblListingDefault th/td`. Especially in some titles e.g. "Balace" on `validate.php`