Highlight the password at reset message sent.
The password is highlighted in red and the menu where change it in green.

There have been adjustments to the parameter replacement of translated texts, the `$` is removed since complicated to write the translated text (rather than `{$ name}` now is used `{name}`) and this allows use new `i18n()` function.