manage_orderable_products: It takes a long time to turn on many products.
When a provider has a lot products (eg over 560) when provider is select it may cause the browser detects excessive time execution and ask *"A script on this page may be busy... continue?"*

Patch proposed optimizes the code and fixes a case where `$('.loadSpinner')` is not displayed when the browser is working.
Note:The main improvement is due to usage of new array `gdatesIsPast[]` that stores if a date is in the past or not.