[PATCH] Remove extra space in email headers
With this extra space, the malformed Cc or Bcc header was considered as the
first line of the body of the email, a new line was inserted before it. This
new line was preventing the following headers (including Content-Type) to be
interpreted as headers and the HTML email was displayed in plain text.
(text by ricola)

Reported by Josu in (https://lists.riseup.net/www/info/aixada-dev)[Aixada-Dev] and tested by ricola.

Please merge, thanks.