The controller should check the order status.
"manage_orders.php" allows close an order if it is open when the list orders is displayed. But what happens if another user has closed the order after displaying the list? Yes... the order will be closed twice, and activating provider products fail!

This patch proposes some controls on the server to prevent actions on orders that no longer have the state shown in the list or in the distribute page.