Use ajaxQueue to work on hosting with few resources.
To work on hosting with few resources (memory) it is recommended to
sequence ajax requests, this avoids simulating requests from a page.

WHY?: We and other cooperative using the same hosting we have seen constant error appeared when loading the Aixada pages. Consulted with the technical service showed the log where says that not can allocate memory, and they recommended that we use a business-pro plan. Avoiding concurrent simultaneous ajax requests the problem disappears completely!