Activate products: maintain the width and height of the cells on hover
The problem was that sometimes the emergent menus disappeared when activated, this was due to the cell size change produced by the 'mouseleave' event.

A color frame of 1px is proposed. It uses the background color and changes color on hover. It also prevents going the font from bold to normal. In this way the cells not change their size on hover.