Use $_SESSION instead of encrypted cookie.
The sensitive information is only saved in $ _SESSION and in this way it is avoided in the client cotens sensitive data even though it was encrypted.

On the other hand when you stop using mcrypt Aixada no longer fails for any page using PHP7.1 or higher.

This is a step to allow PHP7 to be used.