Go to *login.php* when not logged by controller
User can continue on a Aixada page tried to make requests, but in another tab the session is logged out. Therefore, all controllers are responding to requests with error "Not logged in".

It is proposed to show a message and go to `login.php`.
(new method `AixadaNotLoggedInMsg` in `jquery.aixadaUtilities.js` to display this message that is depending on the language)