Allow to connect mySql to a different port.
It is proposed to allow configure `$db_host` using the syntax `$db_host='myHost:9999'` to indicate a different port from the default.

This is useful when developing, allows use different bd systems on the same host, each one in different ports.

NOTE: `$db_type` is removed from `config.php.sample` since it is not used.