BUG: Cumulative order of the UF is lost if after making a normal order.
The error was introduced by me in #239

Since there are no longer two tabs on the shopping cart, one for the date and another for the cumulative ones, there is no more double information in oper=commit of the order of the date and the preoder=true. Currently cumulatives are also worked as of date.

I did not see this `or date_for_order='1234-01-23'` in this `DELETE`, I'm sorry.