Fix issue #303
If a cart is empty it is not possible to validate, it remains in the UF and there is no way to remove.

It is proposed to activate the validation process, in fact `validation_cart_manager` suppresses a empty cart.
Also in the other validation way `direct validate order()` of `utilities/orders.php` remove empty carts of date for shop.
And on validation screen `validate.php`, now validate button is enabled for empty carts.